# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Solstein API
  version: "1"
  summary: Podcast RSS proxy with optional regional exits and ad removal.
  description: |
    Solstein's HTTP interface: the feed and episode URLs podcast clients use,
    the prefix subscribe route, the feed management API and the health check.
    How the features behind it work is described in `docs/` (start at
    `docs/README.md`).

    **Access.** Every route except `/api/health` is refused with `403` for
    clients outside `allowed_client_networks` (when set). On top of that:
    - The prefix route and the feed API need the subscribe token
      (`auth_token` in `config.json`).
    - Feed and episode URLs need the `sig` query parameter that Solstein
      wrote into them: an HMAC over the exact path. Clients get these URLs
      from the rendered feed or from the API's `feed_url`; they are never
      built by hand.
    - With `disable_auth: true`, neither the token nor signatures are
      checked.

    **Errors** are JSON objects with a single `error` message. Internal
    details go to Solstein's log, never into responses.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: "{externalURL}"
    variables:
      externalURL:
        default: http://localhost:8080
        description: Solstein's `external_url`.
...
paths:
...
  /api/v1/feeds/{feedID}:
    parameters:
      - $ref: "#/components/parameters/FeedID"
    get:
      tags: [Feed API]
      summary: Get a feed
      operationId: getFeedSettings
      responses:
        "200":
          description: The feed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Feed API]
      summary: Change a feed's settings
      description: |
        Changes the settings given; omitted fields are left alone, and an
        empty value (`""`, `[]`, `0`) clears the override so the global
        setting applies. Cached episodes made with settings that no longer
        apply (another exit, delivery mode, region-diff settings) are cleared
        at once and prepared again on their next play; failed episodes whose
        settings changed are retried.
      operationId: updateFeed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedSettings"
            examples:
              regionDiffOn:
                value:
                  region_diff: "on"
              backToGlobal:
                value:
                  exit: ""
                  region_diff: ""
      responses:
        "200":
          description: The feed with its new settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
        "400":
          description: Invalid body or settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Feed API]
      summary: Remove a feed
      description: Removes the feed, its episodes and its cached audio.
      operationId: deleteFeed
      responses:
        "204":
          description: Removed.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
security:
  - bearerToken: []
  - queryToken: []
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: "The subscribe token (`auth_token`) as `Authorization: Bearer <token>`."
    queryToken:
      type: apiKey
      in: query
      name: token
      description: The subscribe token as `?token=`, for a quick look in a browser. The bearer header wins when both are given.
  parameters:
    FeedID:
      name: feedID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: Missing or wrong subscribe token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No such feed or episode.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Something failed inside Solstein; the details are in its log.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          examples: ["Feed not found."]
    FeedSettings:
      type: object
      description: |
        Per-feed settings. Each is optional; an empty value means the global
        setting from `config.json` applies.
      properties:
        exit:
          type: string
          description: The exit the feed's polls, downloads and streams go through; empty means `default_exit`.
          examples: [norway, direct]
        delivery_mode:
          type: string
          enum: ["", cache, stream, original]
          description: How audio is delivered when no processor handles the feed.
        poll_interval_minutes:
          type: integer
          minimum: 0
          description: Minutes between polls; 0 means `poll_interval_minutes`.
        region_diff:
          type: string
          enum: ["", "on", "off"]
          description: Region diff for this feed; empty follows `region_diff.enabled`. `on` needs region diff to be running.
        region_diff_exits:
          type: array
          description: The feed's own pair of exits, home region first; empty uses `region_diff.exits`.
          items:
            type: string
          maxItems: 2
          examples: [[norway, sweden]]
        region_diff_on_failure:
          type: string
          enum: ["", publish, hide]
          description: What happens to an episode that can't be processed; empty follows `region_diff.on_failure`.
        region_diff_trim_break_markers:
          type: string
          enum: ["", "on", "off"]
          description: Also remove break markers that can be cut cleanly; empty follows `region_diff.trim_break_markers`.
        region_diff_compare_by_audio:
          type: string
          enum: ["", "on", "off"]
          description: Compare downloads that share no MP3 frame (the host re-encodes around its ads) by their loudness over time; empty follows `region_diff.compare_by_audio` (on by default).
        prepare_ahead:
          type: string
          enum: ["", "on", "off"]
          description: >-
            Prepare every episode before any client asks for it, and keep it out of the feed until its
            file is ready (the backlog included); empty follows `prepare_ahead`.
    Feed:
      description: A subscribed feed, its settings and its state.
      allOf:
        - $ref: "#/components/schemas/FeedSettings"
        - type: object
          required: [id, source_url, feed_url, delivery_mode_in_use, region_diff_in_use, prepare_ahead_in_use]
          properties:
            id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            source_url:
              type: string
              format: uri
              description: The source feed, normalised.
            title:
              type: string
            last_polled_at:
              type: [string, "null"]
              format: date-time
            last_success_at:
              type: [string, "null"]
              format: date-time
            last_error:
              type: string
              description: Why the last poll failed; empty after a successful one.
            delivery_mode_in_use:
              type: string
              enum: [cache, stream, original]
              description: The delivery mode after the global setting is applied.
            region_diff_in_use:
              type: boolean
              description: Whether region diff processes the feed's episodes, after the global and per-feed settings are combined.
            prepare_ahead_in_use:
              type: boolean
              description: Whether the feed's episodes are all prepared before a client asks, after the global and per-feed settings are combined.
            feed_url:
              type: string
              format: uri
              description: The signed feed URL to subscribe to in a podcast client.